splits CardinalityEstimationWorker out of RequestBasedCardinalityEstimator
